requirements.txt
https://github.com/pyrogram/pyrogram/archive/asyncio-dev.zip
requests
tgcrypto
hurry.filesize